requirements.txt
click==7.1.2
Flask==1.1.2
PyMySQL==1.0.2
python-dotenv==0.17.0
ipython>=7.16.1
ipdb==0.13.7
flake8==3.9.1